moving addProcedure to the respective classes
