+++ splice_machine/src/main/java/com/splicemachine/derby/procedures/GenericSpliceFunctions.java
@@ -12,16 +12,38 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.splicemachine.derby.utils;
+package com.splicemachine.derby.procedures;
 
 import com.splicemachine.EngineDriver;
+import com.splicemachine.db.catalog.types.RoutineAliasInfo;
+import com.splicemachine.db.iapi.error.StandardException;
+import com.splicemachine.db.iapi.types.DataTypeDescriptor;
+import com.splicemachine.db.impl.sql.catalog.Procedure;
 import com.splicemachine.primitives.Bytes;
+
+import java.sql.Types;
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * @author Scott Fines
  *         Date: 7/31/14
  */
 public class GenericSpliceFunctions {
+    public static List<Procedure> getProcedures() throws StandardException {
+        return Arrays.asList(
+                //
+                // General functions
+                //
+                Procedure.newBuilder().name("LONG_UUID")
+                        .numOutputParams(1)
+                        .numResultSets(0)
+                        .sqlControl(RoutineAliasInfo.NO_SQL)
+                        .returnType(DataTypeDescriptor.getCatalogType(Types.BIGINT))
+                        .isDeterministic(false).ownerClass(GenericSpliceFunctions.class.getCanonicalName())
+                        .build()
+        );
+    }
 
     private GenericSpliceFunctions(){} //don't instantiate utility class
