+++ splice_machine/src/test/java/com/splicemachine/derby/utils/TimestampProceduresIT.java
@@ -17,7 +17,6 @@ package com.splicemachine.derby.utils;
 import java.sql.CallableStatement;
 import java.sql.ResultSet;
 
-import com.splicemachine.test.SerialTest;
 import org.apache.commons.dbutils.DbUtils;
 import org.junit.*;
 import org.junit.rules.RuleChain;
@@ -31,9 +30,9 @@ import com.splicemachine.derby.test.framework.SpliceWatcher;
  * Integration tests for TimestampAdmin.
  */
 @Ignore("-sf- needs to be re-implemented in an architecture-independent way, but I don't want to let that stop merging")
-public class TimestampAdminIT {
+public class TimestampProceduresIT {
     protected static SpliceWatcher spliceClassWatcher = new SpliceWatcher();
-    public static final String CLASS_NAME = TimestampAdminIT.class.getSimpleName().toUpperCase();
+    public static final String CLASS_NAME = TimestampProceduresIT.class.getSimpleName().toUpperCase();
     protected static SpliceTableWatcher spliceTableWatcher = new SpliceTableWatcher("TEST1",CLASS_NAME,"(a int)");
     protected static SpliceSchemaWatcher spliceSchemaWatcher = new SpliceSchemaWatcher(CLASS_NAME);
